+++ src/app/main/lab-graph/lab-graph.component.css
@@ -1,0 +1,28 @@
+.lab-graph { 
+    padding: 30px;
+}
+
+.lab-graph .lab-graph-header { 
+    position: relative;
+    padding-bottom: 15px;
+}
+
+.lab-graph .lab-graph-header h2 { 
+    text-align: center;
+    margin-bottom: 5px;
+}
+.lab-graph .lab-graph-header h4 { 
+    margin: 10px 0;
+}
+
+.lab-graph .lab-graph-header img.back-arrow { 
+    position: absolute;
+    top: 3px;
+    left: -20px;
+}
+
+.lab-graph .lab-graph-header img.info { 
+    position: absolute;
+    top: 3px;
+    right: -20px;
+}
