+++ src/app/main/lab-graph/lab-graph.component.html
@@ -1,0 +1,15 @@
+<div class="lab-graph">
+    <div class="lab-graph-header">
+        <h2>My GFR Results</h2>
+        <h4>GFR Tests how well your kidneys work</h4>
+        <img class="back-arrow" src="../../../assets/svg/arrow_back_blue.svg" (click)="goBack()" />
+        <img class="info" src="../../../assets/svg/info_blue.svg"/>
+
+    </div>
+    <div class="lab-graph-chart">
+        <canvas baseChart height="200" [datasets]="lineChartData" [labels]="lineChartLabels"
+            [options]="lineChartOptions" [colors]="lineChartColors" [legend]="lineChartLegend"
+            [chartType]="lineChartType" [plugins]="lineChartPlugins">
+        </canvas>
+    </div>
+</div>
