+++ src/app/main/lab-results/lab-results.component.html
@@ -1,0 +1,80 @@
+<div class="lab-results-container">
+    <h2>
+        <img src="../../../assets/svg/beaker.svg">
+        Lab Results
+    </h2>
+    <ul class="lab-result-items">
+        <li routerLink="/lab-graph">
+            <div>
+                <h3 class="lab-result-title">GFR Blood Test</h3>
+                <span class="lab-result-details-1"></span>
+                <span class="lab-result-details lab-result-details-2">September 26, 2020</span>
+                <span class="lab-result-details lab-result-details-3"></span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li routerLink="/lab-graph">
+            <div>
+                <h3 class="lab-result-title">Test Name 2</h3>
+                <span class="lab-result-details-1"></span>
+                <span class="lab-result-details lab-result-details-2">One line of details</span>
+                <span class="lab-result-details lab-result-details-3">September 18, 2020</span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li>
+            <div>
+                <h3 class="lab-result-title">Test Name 3</h3>
+                <span class="lab-result-details-1">September 18, 2020</span>
+                <span class="lab-result-details lab-result-details-2">One line of details</span>
+                <span class="lab-result-details lab-result-details-3"></span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li>
+            <div>
+                <h3 class="lab-result-title">Test Name 4</h3>
+                <span class="lab-result-details-1">September 18, 2020</span>
+                <span class="lab-result-details lab-result-details-2">Two lines of</span>
+                <span class="lab-result-details lab-result-details-3">Details</span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li>
+            <div>
+                <h3 class="lab-result-title">Test Name 5</h3>
+                <span class="lab-result-details-1"></span>
+                <span class="lab-result-details lab-result-details-2">August 3, 2020</span>
+                <span class="lab-result-details lab-result-details-3"></span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li>
+            <div>
+                <h3 class="lab-result-title">Test Name 6</h3>
+                <span class="lab-result-details-1"></span>
+                <span class="lab-result-details lab-result-details-2">September 26, 2020</span>
+                <span class="lab-result-details lab-result-details-3">September 26, 2020</span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li>
+            <div>
+                <h3 class="lab-result-title">GFR Blood Test</h3>
+                <span class="lab-result-details-1">September 26, 2020</span>
+                <span class="lab-result-details lab-result-details-2">September 26, 2020</span>
+                <span class="lab-result-details lab-result-details-3">September 26, 2020</span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+        <li>
+            <div>
+                <h3 class="lab-result-title">GFR Blood Test</h3>
+                <span class="lab-result-details-1">September 26, 2020</span>
+                <span class="lab-result-details lab-result-details-2"></span>
+                <span class="lab-result-details lab-result-details-3"></span>
+            </div>
+            <span class="lab-result-more-icon">&gt;</span>
+        </li>
+    </ul>
+</div>
