+++ src/app/main/lab-results/lab-results.component.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'lab-results',
+  templateUrl: './lab-results.component.html',
+  styleUrls: ['./lab-results.component.css']
+})
+export class LabResultsComponent implements OnInit {
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+}
