+++ src/app/main/status/status.component.css
@@ -115,89 +115,3 @@
     font: 12px Arial, sans-serif;
     vertical-align: middle;
 }
-
-.lab-results-container>h2>img {
-    height: 30px;
-    float: left;
-    padding-right: 10px;
-}
-
-ul.lab-result-items {
-    list-style-type: none;
-    padding: 0;
-}
-
-ul.lab-result-items li {
-    background-color: #F1F9FF;
-    position: relative;
-    height: 90px;
-    float: left;
-    clear: both;
-    width: 100%;
-    margin-bottom: 20px;
-}
-
-ul.lab-result-items li h3 {
-    margin: 0;
-}
-
-ul.lab-result-items li>div {
-    padding: 15px;
-    padding-top: 12px;
-}
-
-ul.lab-result-items li .lab-result-title {
-    font-weight: 500;
-    padding-bottom: 5px;
-}
-
-ul.lab-result-items li .lab-result-details {
-    color: #409FFF;
-    clear: both;
-    float: left;
-    font-size: .8em;
-}
-
-ul.lab-result-items li .lab-result-more-icon {
-    position: absolute;
-    right: 20px;
-    top: 35%;
-    color: #409FFF;
-    font-size: 1.5em;
-}
-
-ul.lab-result-items li .lab-result-details-1 {
-    position: absolute;
-    top: 15px;
-    right: 50px;
-    font-size: .8em;
-}
-
-.lab-graph { 
-    padding: 30px;
-}
-
-.lab-graph .lab-graph-header { 
-    position: relative;
-    padding-bottom: 15px;
-}
-
-.lab-graph .lab-graph-header h2 { 
-    text-align: center;
-    margin-bottom: 5px;
-}
-.lab-graph .lab-graph-header h4 { 
-    margin: 10px 0;
-}
-
-.lab-graph .lab-graph-header img.back-arrow { 
-    position: absolute;
-    top: 3px;
-    left: -20px;
-}
-
-.lab-graph .lab-graph-header img.info { 
-    position: absolute;
-    top: 3px;
-    right: -20px;
-}
